@@ -1,18 +1,21 @@
 source 'https://rubygems.org'
 
-ruby '2.1.4'
+ruby '2.3.3'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.1'
+gem 'rails', '5.1.4'
+# Use Puma as the app server
+gem 'puma', '~> 3.7'
 # Use postgresql as the database for Active Record
 gem 'pg'
-gem 'sass-rails', '~> 4.0.3'
+gem 'sass-rails', '~> 5.0'
 gem 'uglifier', '>= 1.3.0'
-gem 'coffee-rails', '~> 4.0.0'
+gem 'coffee-rails', '~> 4.2'
 
 gem 'jquery-rails'
-gem 'jbuilder', '~> 2.0'
+gem 'jbuilder', '~> 2.5'
 gem 'sdoc', '~> 0.4.0',          group: :doc
+gem 'responders', '~> 2.0'
 
 gem 'rake'
 gem 'simple_form'
@@ -37,3 +40,5 @@ group :development, :test do
   gem 'capybara'
   gem 'factory_girl_rails'
 end
+
+gem 'listen', '>= 3.0.5', '< 3.2', group: :development
