@@ -1,4 +1,4 @@
-class Block < ActiveRecord::Base
+class Block < ApplicationRecord
   has_many :cards, dependent: :destroy
   belongs_to :user
 
