@@ -1,8 +1,8 @@
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   has_many :cards, dependent: :destroy
   has_many :blocks, dependent: :destroy
   has_many :authentications, dependent: :destroy
-  belongs_to :current_block, class_name: 'Block'
+  belongs_to :current_block, class_name: 'Block', optional: true
   before_create :set_default_locale
   before_validation :set_default_locale, on: :create
 
