@@ -1,4 +1,4 @@
-class AddUserIdToCards < ActiveRecord::Migration
+class AddUserIdToCards < ActiveRecord::Migration[5.1]
   def change
     add_column :cards, :user_id, :integer
   end
