@@ -1,4 +1,4 @@
-class SorceryCore < ActiveRecord::Migration
+class SorceryCore < ActiveRecord::Migration[5.1]
   def change
     change_column :users, :email, :string, null: false
     rename_column :users, :password, :crypted_password
