@@ -1,4 +1,4 @@
-class SorceryRememberMe < ActiveRecord::Migration
+class SorceryRememberMe < ActiveRecord::Migration[5.1]
   def change
     add_column :users, :remember_me_token, :string, :default => nil
     add_column :users, :remember_me_token_expires_at, :datetime, :default => nil
