@@ -1,4 +1,4 @@
-class SorceryResetPassword < ActiveRecord::Migration
+class SorceryResetPassword < ActiveRecord::Migration[5.1]
   def change
     add_column :users, :reset_password_token, :string, :default => nil
     add_column :users, :reset_password_token_expires_at, :datetime, :default => nil
