@@ -1,4 +1,4 @@
-class SorceryExternal < ActiveRecord::Migration
+class SorceryExternal < ActiveRecord::Migration[5.1]
   def change
     create_table :authentications do |t|
       t.integer :user_id, null: false
