@@ -1,4 +1,4 @@
-class ChangeEmailTypeFromUsers < ActiveRecord::Migration
+class ChangeEmailTypeFromUsers < ActiveRecord::Migration[5.1]
   def change
     change_column :users, :email, :string, null: true
   end
