@@ -1,4 +1,4 @@
-class AddImageToCards < ActiveRecord::Migration
+class AddImageToCards < ActiveRecord::Migration[5.1]
   def change
     add_column :cards, :image, :string
   end
