@@ -1,4 +1,4 @@
-class CreateBlocks < ActiveRecord::Migration
+class CreateBlocks < ActiveRecord::Migration[5.1]
   def change
     create_table :blocks do |t|
       t.string :title, null: false
