@@ -1,4 +1,4 @@
-class RenameCurrentBlockToCurrentBlockIdFromUsers < ActiveRecord::Migration
+class RenameCurrentBlockToCurrentBlockIdFromUsers < ActiveRecord::Migration[5.1]
   def change
     rename_column :users, :current_block, :current_block_id
   end
