@@ -1,4 +1,4 @@
-class AddReviewStepToCards < ActiveRecord::Migration
+class AddReviewStepToCards < ActiveRecord::Migration[5.1]
   def change
     add_column :cards, :review_step, :integer
   end
