@@ -1,4 +1,4 @@
-class AddLocaleToUsers < ActiveRecord::Migration
+class AddLocaleToUsers < ActiveRecord::Migration[5.1]
   def change
     add_column :users, :locale, :string
   end
