@@ -1,4 +1,4 @@
-class ChangeLocaleNullForUsers < ActiveRecord::Migration
+class ChangeLocaleNullForUsers < ActiveRecord::Migration[5.1]
   def change
     change_column :users, :locale, :string, null: false
   end
