@@ -1,4 +1,4 @@
-class DeleteQualityFromCards < ActiveRecord::Migration
+class DeleteQualityFromCards < ActiveRecord::Migration[5.1]
   def change
     remove_column :cards, :quality
   end
