@@ -1,4 +1,4 @@
-class AddQualityToCards < ActiveRecord::Migration
+class AddQualityToCards < ActiveRecord::Migration[5.1]
   def change
     add_column :cards, :quality, :integer, null: false, default: 5
   end
