@@ -1,6 +1,8 @@
-class Dashboard::UserSessionsController < Dashboard::BaseController
-  def destroy
-    logout
-    redirect_to login_path, notice: t(:log_out_is_successful_notice)
+module Dashboard
+  class UserSessionsController < BaseController
+    def destroy
+      logout
+      redirect_to login_path, notice: t(:log_out_is_successful_notice)
+    end
   end
 end
