@@ -1,12 +1,14 @@
-class Dashboard::UsersController < Dashboard::BaseController
-  def destroy
-    current_user.destroy
-    redirect_to login_path, notice: 'Пользователь успешно удален.'
-  end
+module Dashboard
+  class UsersController < BaseController
+    def destroy
+      current_user.destroy
+      redirect_to login_path, notice: 'Пользователь успешно удален.'
+    end
 
-  private
+    private
 
-  def user_params
-    params.require(:user).permit(:email, :password, :password_confirmation)
+    def user_params
+      params.require(:user).permit(:email, :password, :password_confirmation)
+    end
   end
 end
