@@ -1,2 +1,4 @@
-class Home::BaseController < ApplicationController
+module Home
+  class BaseController < ApplicationController
+  end
 end
