@@ -1,26 +1,29 @@
-class Home::UsersController < Home::BaseController
+module Home
+  class UsersController < BaseController
+    respond_to :html
 
-  def new
-    if current_user
-      redirect_to root_path
-    else
-      @user = User.new
+    def new
+      if current_user
+        redirect_to root_path
+      else
+        @user = User.new
+      end
     end
-  end
 
-  def create
-    @user = User.new(user_params)
-    if @user.save
-      auto_login(@user)
-      redirect_to root_path, notice: t(:user_created_successfully_notice)
-    else
-      respond_with @user
+    def create
+      @user = User.new(user_params)
+      if @user.save
+        auto_login(@user)
+        redirect_to root_path, notice: t(:user_created_successfully_notice)
+      else
+        respond_with @user
+      end
     end
-  end
 
-  private
+    private
 
-  def user_params
-    params.require(:user).permit(:email, :password, :password_confirmation)
+    def user_params
+      params.require(:user).permit(:email, :password, :password_confirmation)
+    end
   end
 end
