@@ -7,7 +7,7 @@ module TrainerHelper
 
   def check_review_card(card, translate, number)
     number.times {
-      put :review_card, { card_id: card.id, user_translation: translate }
+      put :review_card, params: { card_id: card.id, user_translation: translate }
     }
     Card.find(card.id)
   end
